refactor: questionGenerator repository
