--- services/questionService/questionGenerator/test_case_output_generator/__init__.py
@@ -1,3 +1,0 @@
-from .main import generate_outputs
-
-__all__ = ["generate_outputs"]
