--- services/questionService/questionGenerator/utils/__init__.py
@@ -1,3 +1,0 @@
-from .convert_to_json_lines import convert_to_json_lines
-
-__all__ = ["convert_to_json_lines"]
